Allow control of visibility of URI scheme in self-links.
Following on from a discussion about the removal of the URI scheme
from self-links I have added an option to show it for all allowable
schemes (except for 'mailto') if needed. I'm not 100% happy with
leaving mailto out of the bag at the moment as this gives an
inconsistent implementation.