Square brackets can be used to enforce line code tag rendering
Touches the issues outlined in #129. With the way PHP-Textile parser
works, having code tags directly nested within span tags will not be
possible, not without reworking the whole parser, or introducing error
prone regexes to match code tags, but in the meantime the bracket syntax
can be used to enforce code rendering.

Closes #129