Fix PHP 8.4 compatibility
Fixes PHP 8.4 compatibility in backwards compatible
fashion. Also updates testing suite to be
compatible with PHP 8.4, but doesn't yet
use it in CI as the run time, but it can
be executed locally.

PHP 8.4 is not released yet, and compatible
xdebug version is not yet available in PECL
repository.

Fixes #227